requirements.txt
annotated-types==0.7.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
contourpy>=1.3.2
cycler==0.12.1
dnspython==2.8.0
filelock==3.13.1
fonttools==4.60.0
fsspec==2024.6.1
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.9
MarkupSafe==2.1.5
matplotlib==3.10.6
minio==7.2.18
mpmath==1.3.0
networkx==3.3
numpy>=2.0.0
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==12.0.0
polars==1.33.1
psutil==7.1.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.3
pydantic-core==2.41.4
pymongo==4.15.3
pyparsing==3.2.4
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.5
scipy==1.15.3
setuptools==63.2.0
six==1.17.0
sympy==1.13.3
torch==2.9.1
torchvision==0.24.1
typing-extensions==4.15.0
typing-inspection==0.4.2
ultralytics==8.3.202
ultralytics-thop==2.0.17
urllib3==2.5.0
watchdog==6.0.0